+++ src/core/api/types.ts
@@ -1,4 +1,4 @@
 export type mesh = { 
     vertices: Float32Array;
-    indices: Uint16Array;
+    indices: Uint32Array;
 }
